Fix warnings
git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/turbine/trunk@5463 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ code/src/tcl/blob/blob.c
@@ -36,7 +36,7 @@ SwiftBlob_pointer_as_integer(void* p)
 double*
 SwiftBlob_cast_to_pointer(int i)
 {
-  return (double*) i;
+  return (double*) (size_t) i;
 }
 
 double
